IMAP-370 Solve inheritance bugg on request handling
