+++ protocols/imap/src/main/java/org/apache/james/imap/decode/parser/MoveCommandParser.java
@@ -8,12 +8,13 @@ import org.apache.james.imap.message.request.MoveRequest;
 /**
  * Parse MOVE commands
  */
-public class MoveCommandParser extends CopyCommandParser {
+public class MoveCommandParser extends AbstractMessageRangeCommandParser {
 	
 	public MoveCommandParser() {
 		super(ImapCommand.selectedStateCommand(ImapConstants.MOVE_COMMAND_NAME));
 	}
-	
+
+	@Override
 	protected MoveRequest createRequest(ImapCommand command, String tag,
 			boolean useUids, IdRange[] idSet, String mailboxName) {
 		return new MoveRequest(command, idSet, mailboxName, useUids, tag);
